+++ src/LiveComponent/src/Attribute/AsLiveComponent.php
@@ -22,7 +22,7 @@ use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 final class AsLiveComponent extends AsTwigComponent
 {
     public function __construct(
-        string $name,
+        ?string $name = null,
         ?string $template = null,
         private ?string $defaultAction = null,
         bool $exposePublicProps = true,