+++ ux.symfony.com/src/Twig/FoodVote.php
@@ -11,8 +11,8 @@ use Symfony\UX\LiveComponent\Attribute\LiveArg;
 use Symfony\UX\LiveComponent\Attribute\LiveProp;
 use Symfony\UX\LiveComponent\DefaultActionTrait;
 
-#[AsLiveComponent('food_vote')]
-class FoodVoteComponent extends AbstractController
+#[AsLiveComponent()]
+class FoodVote extends AbstractController
 {
     use DefaultActionTrait;